requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
motor==3.3.2
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
aiohttp==3.9.3
httpx==0.26.0

# Testes
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# Desenvolvimento
black==24.1.1
isort==5.13.2
flake8==7.0.0
flake8-docstrings==1.7.0
mypy==1.8.0
pre-commit==3.6.0 